fix(F-495): replace client side rendering
