+++ frontend/src/app/[locale]/(main)/admin/components/Reviews.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { ChevronDown, Star } from 'lucide-react';
-import { useTranslations } from 'next-intl';
+import { useLocale, useTranslations } from 'next-intl';
 import { useEffect, useState } from 'react';
 
 import { Button } from '@/components/ui/Button';
@@ -19,12 +19,14 @@ import {
 } from '@/components/ui/Select';
 import { useRouter } from 'next/navigation';
 import EmptyListComponent from '@/components/common/EmptyListComponent';
+import { formattedDate } from '@/lib/utils';
 
 export default function Reviews({ ratingStatistics, reviews, pagination }: ReviewsPaginated) {
   const limit = pagination?.pageSize;
   const router = useRouter();
   const t = useTranslations('Admin');
   const tCommon = useTranslations('Common');
+  const locale = useLocale();
   const [visibleCount, setVisibleCount] = useState(limit);
   const [pageNumber, setPageNumber] = useState(pagination?.currentPage);
   const [isLoading, setIsLoading] = useState(false);
@@ -151,9 +153,7 @@ export default function Reviews({ ratingStatistics, reviews, pagination }: Revie
                   ))}
                 </div>
                 <div className="text-sm text-bw-70 mb-2">{review.comment}</div>
-                <div className="text-sm text-bw-40">
-                  {new Date(review.date).toLocaleDateString()}
-                </div>
+                <div className="text-sm text-bw-40">{formattedDate(review.date, locale)}</div>
               </div>
             ))}
           </div>
