+++ frontend/src/app/[locale]/(standalone)/simulation/[id]/components/SimulationHeader.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { User } from 'lucide-react';
 import { Avatar, AvatarImage, AvatarFallback } from '@/components/ui/Avatar';
 import { SimulationHeaderProps } from '@/interfaces/SimulationHeaderProps';
