+++ frontend/src/components/common/EmptyListComponent.tsx
@@ -1,9 +1,9 @@
 import Image from 'next/image';
-import { useTranslations } from 'next-intl';
 import { EmptyListComponentProps } from '@/interfaces/EmptyListComponentProps';
+import { getTranslations } from 'next-intl/server';
 
-export default function EmptyListComponent({ itemType }: EmptyListComponentProps) {
-  const t = useTranslations('Common.emptyList');
+export default async function EmptyListComponent({ itemType }: EmptyListComponentProps) {
+  const t = await getTranslations('Common.emptyList');
   return (
     <div className="flex flex-col items-center justify-center w-full border border-bw-20 rounded-lg bg-transparent gap-1 p-8 text-center">
       <Image
